gcp/requirements.txt
tensorflow==2.13.0
google-cloud-storage==2.10.0
Pillow==10.0.0
numpy==1.23.5